Move build settings test utilities to test fixtures
