+++ subprojects/build-adapter-xcode/build-adapter-xcode.gradle
@@ -8,6 +8,7 @@ plugins {
 	id 'dev.gradleplugins.documentation.api-reference-module'
 	id 'nokeedocs.sample-templates'
 	id 'antlr'
+	id 'java-test-fixtures'
 }
 
 dependencies {
@@ -27,6 +28,8 @@ dependencies {
 	implementation project(':xcodeIdeKit')
 	implementation "org.antlr:antlr4:${antlrVersion}"
 	antlr "org.antlr:antlr4:${antlrVersion}"
+
+	testFixturesApi project(":internalTesting")
 }
 tasks.named('generateGrammarSource', AntlrTask) { task ->
 	task.arguments += ['-visitor']
@@ -51,6 +54,7 @@ test {
 		implementation templates(project(':xcodeIdeKit'))
 		implementation testFixtures(project(':coreExec'))
 		implementation "com.google.jimfs:jimfs:${jimfsVersion}"
+		implementation testFixtures(project)
 	}
 }
 
@@ -58,6 +62,7 @@ integrationTest {
 	dependencies {
 		implementation 'dev.nokee:templates'
 		implementation templates(project(':xcodeIdeKit'))
+		implementation testFixtures(project)
 	}
 	testTasks.configureEach {
 		maxParallelForks = Runtime.runtime.availableProcessors()
@@ -80,6 +85,7 @@ functionalTest {
 		implementation templates(project)
 		implementation 'com.google.code.gson:gson:2.8.6'
 		implementation project(':coreExec')
+		implementation testFixtures(project)
 	}
 }
