+++ subprojects/build-adapter-xcode/src/test/java/dev/nokee/xcode/buildsettings/DefaultXCBuildSettingLayerTests.java
@@ -23,10 +23,10 @@ import lombok.val;
 import org.junit.jupiter.api.Test;
 
 import static dev.nokee.internal.testing.SerializableMatchers.isSerializable;
-import static dev.nokee.xcode.buildsettings.XCBuildSettingTestUtils.buildSetting;
-import static dev.nokee.xcode.buildsettings.XCBuildSettingTestUtils.buildSettingNamed;
-import static dev.nokee.xcode.buildsettings.XCBuildSettingTestUtils.mapOf;
-import static dev.nokee.xcode.buildsettings.XCBuildSettingTestUtils.nullBuildSetting;
+import static dev.nokee.buildadapter.xcode.testers.XCBuildSettingTestUtils.buildSetting;
+import static dev.nokee.buildadapter.xcode.testers.XCBuildSettingTestUtils.buildSettingNamed;
+import static dev.nokee.buildadapter.xcode.testers.XCBuildSettingTestUtils.mapOf;
+import static dev.nokee.buildadapter.xcode.testers.XCBuildSettingTestUtils.nullBuildSetting;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.equalTo;
